+++ src/main/java/com/nali/list/netmethods/clientmessage/CallPlayerGUI.java
@@ -6,7 +6,7 @@ import static com.nali.small.gui.OpenGUIHelper.callPlayerGUI;
 
 public class CallPlayerGUI
 {
-    public static int ID = 2;
+    public static byte ID = 2;
 
     public static void run(ClientMessage clientmessage)
     {
